Add files via upload
A python module for 2D or 3D feature generation an AI-based Surrogate Reservoir Model (SRM). The permeability field can be generated using a uniform distribution or stochastically by use of the Karhunen Loeve Expansion (KLM). For the KLM expansion, @Ilias Bilionis (2014) python module has been adapted for use in this module.

The first section of the module, and can be altered to specific reservoir and well configurations.